Extract EditRuleDialog in own file
